app: Split client entry point out into a separate file
This will allow us to import the app code from the server side
renderer
